Harden against missed 'end' events
This handles a case where a stream may be ended before being consumed,
and then the `end` event emits as soon as a data handler is added.

Since core streams aggressively de-zalgo by making every event
asynchronous (at great cost), it's common to have this kind of pattern
in stream consumption code:

    function doStuff (stream, callback) {
      let str = ''
      stream.on('data', d => str += d)
      stream.on('end', () => callback(str))
    }

Prior to this patch, that would result in a dangling callback chain.
Now, when an 'end' event listener is added, if an 'end' was previously
emitted _without any listeners attached_, it will re-emit.

Note that this _won't_ protect against this pattern when the stream is
already being piped somewhere, or if 'end' is listened to by some other
event listener.

We could guarantee that every 'end' listener is called exactly once by
removing them after they've been called, and then treating every end as
"silent" for the purposes of adding new consumers.  (That is, go ahead
and re-emit any time an 'end' listener is addded on an ended stream,
knowing that listener would be removed on end emission.)