More caveats and explanation re MP vs node-core
Added a section to the docs outlining the caveats around tee-streams,
and a deeper explanation of why node-core streams are slower than
Minipass streams.

The comment in dfb4ee04d53df6bf7ebee9559fe8b34349316921 about supporting
tee-streams is incorrect, as it turns out.  There is no way to reliably
detect whether a stream should start flowing on pipe() if it has
buffered data, without breaking the fundamental contract of Minipass.

The docs will hopefully help others avoid tee-stream data-loss hazards
in a way that is more in keeping with the contract this library
provides.