Extend Stream instead of EventEmitter
Close #10

A lot of code out there in the wild does `x instanceof Stream` to decide
whether something is a stream, and otherwise, treats it as a generic
Plain Old JavaScript object.

For example, setting `context.body = stream` will cause Koa to print out
something like:

```json
{
  "_events": {
    "error": [
      null,
      null
    ]
  },
  "_eventsCount": 4,
  "pipes": {
    "tail": null,
    "head": null,
    "length": 0
  },
  "buffer": {
    "tail": null,
    "head": null,
    "length": 0
  },
  "writable": true,
  "readable": true
}
```

instead of piping the body through.

The `Stream` class in node is just an EE with a `pipe` method (which
Minipass would override anyway), so it's safe and does not impose any
extra performance hit.