fix: 'readable' not missed, 'drain' not doubled
A complex scenario involving several promises, pipelines, flush streams,
and an internal stream proxy led to an interaction with
npm-registry-fetch, make-fetch-happen, and cacache, where data that
arrived in multiple chunks would end up out of order in the caller.

The root cause seems to be that 'drain' events were being fired
excessively, when nothing had actually been buffered or caused
backpressure.

This patch addresses the problem by ensuring that 'drain' events are
ONLY fired when necessary, because either (a) while in flowing mode a
downstream pipe destination returned 'false' from a write, or (b) in
non-flowing mode, data was not fully consumed in the 'readable' event
emitted during write().

This also makes it so that adding a 'readable' listener after data has
been added to the buffer will emit 'readable' right away, rather than
waiting for a subsequent write, so that 'readable' events are never
missed.