Flush buffer before emitting 'data' during write()
It takes quite a convoluted set of streams to reproduce this error, but
it's possible to end up in a state where there is a chunk in the buffer,
AND we're in a flowing state, during a write() call.  Since the
assumption was that a flowing state means that the buffer must be empty,
this resulted in sending data out of order, breaking the streaming in
make-fetch-happen and npm-registry-fetch, due to how cacache was
proxying writes to a slow tmp file writer, and how make-fetch-happen was
creating a Pipeline to produce the caching response body.

Fix: https://github.com/npm/npm-registry-fetch/issues/23
Re: https://github.com/npm/make-fetch-happen/commit/7f896bef4bbcab73eccb92b5999204b673cdf485