Define event argument types extensibly
It's kind of annoying that Minipass subclasses have to either do a bunch
of `declare` hackery and specify a ton of extra signatures for `emit`
and `on` and `addListener`, or just accept every event being
emitted/handled with `any[]` as the argument list.

Because this changes the default event argument signature from `any[]`
to `unknown[]` (which is safer, but more restrictive), it *may* cause
some disruption.  Even though Minipass v5 is not yet in very wide usage,
and this is arguably a bugfix since the arguments really *are* "unknown"
if not declared, it's important to err on the side of caution with
something like Minipass, as it has the potential to disrupt a lot of
other downstream consumers.