allow symbol as an event type
This was hiding because I have 'skipLibCheck' in most of my packages,
since it makes TS builds so much faster, and it's typically not a
problem, since libs get checked in their own development anyway.  But in
the case of something like Minipass, where it's not *actually*
typescript, but just JavaScript with a hand-written .d.ts, things can go
unnoticed.

Still an open todo to migrate minipass to a TS hybrid build, but in the
meantime, this fixes the problem for node-glob.

Fix: https://github.com/isaacs/node-glob/issues/529
Fix: https://github.com/isaacs/minipass/issues/49
Close: https://github.com/isaacs/minipass/pull/50